+++ kit/src/main/java/io/github/dornol/excelkit/excel/ExcelSheetWriter.java
@@ -510,7 +510,7 @@ public class ExcelSheetWriter<T> {
                 c.minWidth, c.maxWidth, c.fixedWidth, c.dropdownOptions,
                 c.cellColorFunction, c.groupName, c.outlineLevel,
                 c.commentFunction, c.borderStyle, c.locked, c.hidden, c.validation,
-                c.headerFontColor, c.nullValue);
+                c.headerFontColor, c.headerComment, c.nullValue);
     }
 
 }